requirements.txt
requests
aiogram
aiohttp
pytesseract
tox
opencv-python
duckduckgo_search
python-dotenv
youtube_transcript_api
selenium
readability-lxml
bs4
pypdf
python-docx
openpyxl
python-pptx
pyyaml
imaginepy
speechrecognition
pydub